+++ packages/go_router/lib/src/configuration.dart
@@ -8,9 +8,9 @@ import 'package:flutter/widgets.dart';
 import 'configuration.dart';
 import 'logging.dart';
 import 'matching.dart';
+import 'misc/errors.dart';
 import 'path_utils.dart';
 import 'typedefs.dart';
-
 export 'route.dart';
 export 'state.dart';
 
@@ -22,89 +22,97 @@ class RouteConfiguration {
     required this.redirectLimit,
     required this.topRedirect,
     required this.navigatorKey,
-  }) {
+  })  : assert(_debugCheckPath(routes, true)),
+        assert(
+            _debugVerifyNoDuplicatePathParameter(routes, <String, GoRoute>{})),
+        assert(_debugCheckParentNavigatorKeys(
+            routes, <GlobalKey<NavigatorState>>[navigatorKey])) {
     _cacheNameToPath('', routes);
-
     log.info(_debugKnownRoutes());
+  }
 
-    assert(() {
-      for (final RouteBase route in routes) {
-        if (route is GoRoute && !route.path.startsWith('/')) {
+  static bool _debugCheckPath(List<RouteBase> routes, bool isTopLevel) {
+    for (final RouteBase route in routes) {
+      late bool subRouteIsTopLevel;
+      if (route is GoRoute) {
+        if (isTopLevel) {
           assert(route.path.startsWith('/'),
-              'top-level path must start with "/": ${route.path}');
-        } else if (route is ShellRouteBase) {
-          for (final RouteBase route in routes) {
-            if (route is GoRoute) {
-              assert(route.path.startsWith('/'),
-                  'top-level path must start with "/": ${route.path}');
-            }
-          }
+              'top-level path must start with "/": $route');
+        } else {
+          assert(!route.path.startsWith('/') && !route.path.endsWith('/'),
+              'sub-route path may not start or end with /: $route');
         }
+        subRouteIsTopLevel = false;
+      } else if (route is ShellRouteBase) {
+        subRouteIsTopLevel = isTopLevel;
       }
+      _debugCheckPath(route.routes, subRouteIsTopLevel);
+    }
+    return true;
+  }
 
-      // Check that each parentNavigatorKey refers to either a ShellRoute's
-      // navigatorKey or the root navigator key.
-      void checkParentNavigatorKeys(
-          List<RouteBase> routes,
-          List<GlobalKey<NavigatorState>> allowedKeys,
-          Set<GlobalKey<NavigatorState>> allNavigatorKeys) {
-        for (final RouteBase route in routes) {
-          if (route is GoRoute) {
-            final GlobalKey<NavigatorState>? parentKey =
-                route.parentNavigatorKey;
-            if (parentKey != null) {
-              // Verify that the root navigator or a ShellRoute ancestor has a
-              // matching navigator key.
-              assert(
-                  allowedKeys.contains(parentKey),
-                  'parentNavigatorKey $parentKey must refer to'
-                  " an ancestor ShellRoute's navigatorKey or GoRouter's"
-                  ' navigatorKey');
+  // Check that each parentNavigatorKey refers to either a ShellRoute's
+  // navigatorKey or the root navigator key.
+  static bool _debugCheckParentNavigatorKeys(
+      List<RouteBase> routes, List<GlobalKey<NavigatorState>> allowedKeys) {
+    for (final RouteBase route in routes) {
+      if (route is GoRoute) {
+        final GlobalKey<NavigatorState>? parentKey = route.parentNavigatorKey;
+        if (parentKey != null) {
+          // Verify that the root navigator or a ShellRoute ancestor has a
+          // matching navigator key.
+          assert(
+              allowedKeys.contains(parentKey),
+              'parentNavigatorKey $parentKey must refer to'
+              " an ancestor ShellRoute's navigatorKey or GoRouter's"
+              ' navigatorKey');
 
-              allNavigatorKeys.add(parentKey);
-              checkParentNavigatorKeys(
-                route.routes,
-                <GlobalKey<NavigatorState>>[
-                  // Once a parentNavigatorKey is used, only that navigator key
-                  // or keys above it can be used.
-                  ...allowedKeys.sublist(0, allowedKeys.indexOf(parentKey) + 1),
-                ],
-                allNavigatorKeys,
-              );
-            } else {
-              checkParentNavigatorKeys(
-                route.routes,
-                <GlobalKey<NavigatorState>>[
-                  ...allowedKeys,
-                ],
-                allNavigatorKeys,
-              );
-            }
-          } else if (route is ShellRoute && route.navigatorKey != null) {
-            allNavigatorKeys.add(route.navigatorKey);
-            checkParentNavigatorKeys(
-              route.routes,
-              <GlobalKey<NavigatorState>>[
-                ...allowedKeys..add(route.navigatorKey)
-              ],
-              allNavigatorKeys,
-            );
-          } else if (route is StatefulShellRoute) {
-            checkParentNavigatorKeys(
-                route.routes, allowedKeys, allNavigatorKeys);
-          }
+          _debugCheckParentNavigatorKeys(
+            route.routes,
+            <GlobalKey<NavigatorState>>[
+              // Once a parentNavigatorKey is used, only that navigator key
+              // or keys above it can be used.
+              ...allowedKeys.sublist(0, allowedKeys.indexOf(parentKey) + 1),
+            ],
+          );
+        } else {
+          _debugCheckParentNavigatorKeys(
+            route.routes,
+            <GlobalKey<NavigatorState>>[
+              ...allowedKeys,
+            ],
+          );
         }
+      } else if (route is ShellRoute && route.navigatorKey != null) {
+        _debugCheckParentNavigatorKeys(
+          route.routes,
+          <GlobalKey<NavigatorState>>[...allowedKeys..add(route.navigatorKey)],
+        );
+      } else if (route is StatefulShellRoute) {
+        _debugCheckParentNavigatorKeys(route.routes, allowedKeys);
       }
+    }
+    return true;
+  }
 
-      final Set<GlobalKey<NavigatorState>> allNavigatorKeys =
-          <GlobalKey<NavigatorState>>{navigatorKey};
-      checkParentNavigatorKeys(
-          routes, <GlobalKey<NavigatorState>>[navigatorKey], allNavigatorKeys);
-
-      _debugAllNavigatorKeys = allNavigatorKeys;
-
-      return true;
-    }());
+  static bool _debugVerifyNoDuplicatePathParameter(
+      List<RouteBase> routes, Map<String, GoRoute> usedPathParams) {
+    for (final RouteBase route in routes) {
+      if (route is! GoRoute) {
+        continue;
+      }
+      for (final String pathParam in route.pathParams) {
+        if (usedPathParams.containsKey(pathParam)) {
+          final bool sameRoute = usedPathParams[pathParam] == route;
+          throw GoError(
+              "duplicate path parameter, '$pathParam' found in ${sameRoute ? '$route' : '${usedPathParams[pathParam]}, and $route'}");
+        }
+        usedPathParams[pathParam] = route;
+      }
+      _debugVerifyNoDuplicatePathParameter(route.routes, usedPathParams);
+      route.pathParams.forEach(usedPathParams.remove);
+    }
+    return true;
   }
 
   /// The list of top level routes used by [GoRouterDelegate].
@@ -121,7 +129,8 @@ class RouteConfiguration {
 
   final Map<String, String> _nameToPath = <String, String>{};
 
-  late final Set<GlobalKey<NavigatorState>>? _debugAllNavigatorKeys;
+  late final Set<GlobalKey<NavigatorState>> _debugAllNavigatorKeys =
+      _debugAllNavigatorsRecursively(routes);
 
   /// Looks up the url location by a [GoRoute]'s name.
   String namedLocation(
@@ -165,6 +174,25 @@ class RouteConfiguration {
         .toString();
   }
 
+  static Set<GlobalKey<NavigatorState>> _debugAllNavigatorsRecursively(
+      List<RouteBase> routes) {
+    return routes.expand((RouteBase e) {
+      if (e is GoRoute && e.parentNavigatorKey != null) {
+        return <GlobalKey<NavigatorState>>{
+          e.parentNavigatorKey!,
+          ..._debugAllNavigatorsRecursively(e.routes)
+        };
+      } else if (e is ShellRoute) {
+        return <GlobalKey<NavigatorState>>{
+          e.navigatorKey,
+          ..._debugAllNavigatorsRecursively(e.routes)
+        };
+      } else {
+        return _debugAllNavigatorsRecursively(e.routes);
+      }
+    }).toSet();
+  }
+
   /// Validates the branches of a [StatefulShellRoute].
   bool debugValidateStatefulShellBranches(
       StatefulShellRoute shellRoute, List<StatefulShellBranch> branches) {
@@ -180,7 +208,7 @@ class RouteConfiguration {
           'StatefulShellRoute must not uses duplicate defaultLocations for the branches');
 
       assert(
-          _debugAllNavigatorKeys!
+          _debugAllNavigatorKeys
               .intersection(uniqueBranchNavigatorKeys)
               .isEmpty,
           'StatefulShellBranch Navigator key must be unique');