+++ packages/go_router/test/delegate_test.dart
@@ -5,6 +5,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:go_router/go_router.dart';
+import 'package:go_router/src/delegate.dart';
 import 'package:go_router/src/match.dart';
 import 'package:go_router/src/misc/error_screen.dart';
 
@@ -105,10 +106,6 @@ void main() {
       (WidgetTester tester) async {
         final GoRouter goRouter = await createGoRouter(tester);
         expect(goRouter.routerDelegate.matches.matches.length, 1);
-        expect(
-          goRouter.routerDelegate.matches.matches[0].pageKey,
-          null,
-        );
 
         goRouter.push('/a');
         await tester.pumpAndSettle();
@@ -216,47 +213,49 @@ void main() {
   });
 
   group('replace', () {
-    testWidgets(
-      'It should replace the last match with the given one',
-      (WidgetTester tester) async {
-        final GoRouter goRouter = GoRouter(
-          initialLocation: '/',
-          routes: <GoRoute>[
-            GoRoute(path: '/', builder: (_, __) => const SizedBox()),
-            GoRoute(path: '/page-0', builder: (_, __) => const SizedBox()),
-            GoRoute(path: '/page-1', builder: (_, __) => const SizedBox()),
-          ],
-        );
-        await tester.pumpWidget(
-          MaterialApp.router(
-            routerConfig: goRouter,
-          ),
-        );
+    testWidgets('It should replace the last match with the given one',
+        (WidgetTester tester) async {
+      final GoRouter goRouter = GoRouter(
+        initialLocation: '/',
+        routes: <GoRoute>[
+          GoRoute(path: '/', builder: (_, __) => const SizedBox()),
+          GoRoute(path: '/page-0', builder: (_, __) => const SizedBox()),
+          GoRoute(path: '/page-1', builder: (_, __) => const SizedBox()),
+        ],
+      );
+      await tester.pumpWidget(
+        MaterialApp.router(
+          routerConfig: goRouter,
+        ),
+      );
 
-        goRouter.push('/page-0');
+      goRouter.push('/page-0');
 
-        goRouter.routerDelegate.addListener(expectAsync0(() {}));
-        final RouteMatch first = goRouter.routerDelegate.matches.matches.first;
-        final RouteMatch last = goRouter.routerDelegate.matches.last;
-        goRouter.replace('/page-1');
-        expect(goRouter.routerDelegate.matches.matches.length, 2);
-        expect(
-          goRouter.routerDelegate.matches.matches.first,
-          first,
-          reason: 'The first match should still be in the list of matches',
-        );
-        expect(
-          goRouter.routerDelegate.matches.last,
-          isNot(last),
-          reason: 'The last match should have been removed',
-        );
-        expect(
-          goRouter.routerDelegate.matches.last.fullpath,
-          '/page-1',
-          reason: 'The new location should have been pushed',
-        );
-      },
-    );
+      goRouter.routerDelegate.addListener(expectAsync0(() {}));
+      final RouteMatch first = goRouter.routerDelegate.matches.matches.first;
+      final RouteMatch last = goRouter.routerDelegate.matches.last;
+      goRouter.replace('/page-1');
+      expect(goRouter.routerDelegate.matches.matches.length, 2);
+      expect(
+        goRouter.routerDelegate.matches.matches.first,
+        first,
+        reason: 'The first match should still be in the list of matches',
+      );
+      expect(
+        goRouter.routerDelegate.matches.last,
+        isNot(last),
+        reason: 'The last match should have been removed',
+      );
+      expect(
+        (goRouter.routerDelegate.matches.last as ImperativeRouteMatch)
+            .matches
+            .uri
+            .toString(),
+        '/page-1',
+        reason: 'The new location should have been pushed',
+      );
+    });
+
     testWidgets(
       'It should return different pageKey when replace is called',
       (WidgetTester tester) async {
@@ -264,7 +263,7 @@ void main() {
         expect(goRouter.routerDelegate.matches.matches.length, 1);
         expect(
           goRouter.routerDelegate.matches.matches[0].pageKey,
-          null,
+          isNotNull,
         );
 
         goRouter.push('/a');
@@ -331,17 +330,11 @@ void main() {
         );
         expect(
           goRouter.routerDelegate.matches.last,
-          isA<RouteMatch>()
-              .having(
-                (RouteMatch match) => match.fullpath,
-                'match.fullpath',
-                '/page-1',
-              )
-              .having(
-                (RouteMatch match) => (match.route as GoRoute).name,
-                'match.route.name',
-                'page1',
-              ),
+          isA<RouteMatch>().having(
+            (RouteMatch match) => (match.route as GoRoute).name,
+            'match.route.name',
+            'page1',
+          ),
           reason: 'The new location should have been pushed',
         );
       },