+++ packages/pigeon/platform_tests/test_plugin/lib/test_plugin.dart
@@ -2,7 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// TODO(tarrinneal): Add integration tests.
-
-/// A host plugin for Pigeon tests.
-class TestPlugin {}
+// There is intentionally no code here; tests use generated Pigeon APIs
+// directly, as wrapping them in a plugin would just add maintenance burden
+// when changing tests.