+++ packages/go_router/CHANGELOG.md
@@ -1,10 +1,16 @@
-## 5.2.0
+## 5.3.0
 
 - Introduced a new shell route class called `StatefulShellRoute`, to support using separate
   navigators for child routes as well as preserving state in each navigation tree
   (flutter/flutter#99124).
 - Updated documentation for `pageBuilder` and `builder` fields of `ShellRoute`, to more correctly
   describe the meaning of the child argument in the builder functions.
+- Added support for restorationId to ShellRoute (and StatefulShellRoute).
+
+## 5.2.0
+
+- Fixes `GoRouterState.location` and `GoRouterState.param` to return correct value.
+- Cleans up `RouteMatch` and `RouteMatchList` API.
 
 ## 5.1.10